refactor(starknet_l1_provider): extract `L1Provider` into file
Only EXTRACT, no changes except visibility changes inside L1Provider
fields.
