+++ crates/starknet_sequencer_node/src/components.rs
@@ -3,8 +3,9 @@ use starknet_batcher::batcher::{create_batcher, Batcher};
 use starknet_consensus_manager::consensus_manager::ConsensusManager;
 use starknet_gateway::gateway::{create_gateway, Gateway};
 use starknet_http_server::http_server::{create_http_server, HttpServer};
+use starknet_l1_provider::event_identifiers_to_track;
+use starknet_l1_provider::l1_provider::{create_l1_provider, L1Provider};
 use starknet_l1_provider::l1_scraper::L1Scraper;
-use starknet_l1_provider::{create_l1_provider, event_identifiers_to_track, L1Provider};
 use starknet_mempool::communication::{create_mempool, MempoolCommunicationWrapper};
 use starknet_mempool_p2p::create_p2p_propagator_and_runner;
 use starknet_mempool_p2p::propagator::MempoolP2pPropagator;
